# docs/swagger/swagger.yaml
basePath: /
definitions:
  go-micro_internal_activity.CouponRequest:
    properties:
      amount:
        type: number
      coupon_id:
        type: string
      user_id:
        type: string
    type: object
  go-micro_internal_activity.SeckillRequest:
    properties:
      quantity:
        type: integer
      sku_id:
        type: string
      user_id:
        type: string
    type: object
  go-micro_internal_payment.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      order_id:
        type: string
      request_id:
        type: string
    type: object
...
  go-micro_pkg_httpx.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_gateway.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_gateway.Item'
        type: array
      remark:
        type: string
      request_id:
        type: string
      user_id:
        type: string
    type: object
  internal_gateway.Item:
    properties:
      price:
        type: integer
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
  internal_gateway.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  internal_order.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_order.Item'
        type: array
      remark:
        type: string
      request_id:
        type: string
      user_id:
        type: string
    type: object
  internal_order.Item:
    properties:
      price:
        type: integer
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
  internal_payment.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      order_id:
        type: string
      request_id:
        type: string
    type: object
  internal_price.CalculateRequest:
    properties:
      base_price:
        type: number
      coupon_amount:
        type: number
      sku_id:
        type: string
      user_level:
        type: integer
    type: object
  internal_refund.InitiateRequest:
    properties:
      order_id:
        type: string
      reason:
        type: string
      refund_id:
        type: string
      refund_type:
        type: string
    type: object
  internal_refund.RollbackRequest:
    properties:
      order_id:
        type: string
      reason:
        type: string
      refund_id:
        type: string
    type: object
  internal_refund.StatusRequest:
    properties:
      refund_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Go-Micro 网关服务 API
  title: Go-Micro Gateway API
  version: "1.0"
paths:
  /api/v1/activity/coupon:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: issue coupon
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-micro_internal_activity.CouponRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Issue coupon
      tags:
      - Activity
  /api/v1/activity/seckill:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: seckill request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-micro_internal_activity.SeckillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Seckill request
      tags:
      - Activity
  /api/v1/activity/status:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: coupon id
        in: query
        name: coupon_id
        type: string
      - description: sku id
        in: query
        name: sku_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Activity status
      tags:
      - Activity
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_gateway.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 用户登录
      tags:
      - Auth
  /api/v1/order-views/{order_no}:
    get:
      description: |-
        view_status 为聚合层统一展示状态，优先级高于单一服务原始状态。
        view_status 枚举：PENDING、PROCESSING、SUCCESS、FAILED、DEAD、CANCELED、TIMEOUT。
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 业务订单号
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 订单聚合视图
      tags:
      - Order
  /api/v1/orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 订单创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_gateway.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建订单
      tags:
      - Order
  /api/v1/orders/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询订单
      tags:
      - Order
  /api/v1/payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建支付请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-micro_internal_payment.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建支付
      tags:
      - Payment
  /api/v1/payments/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询支付
      tags:
      - Payment
  /api/v1/payments/{id}/failed:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付失败
      tags:
      - Payment
  /api/v1/payments/{id}/success:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付成功
      tags:
      - Payment
  /api/v1/payments/{id}/timeout:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付超时
      tags:
      - Payment
  /api/v1/price/calculate:
    post:
      consumes:
      - application/json
      parameters:
      - description: calculate price
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_price.CalculateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Calculate price
      tags:
      - Price
  /api/v1/price/history:
    get:
      parameters:
      - description: sku id
        in: query
        name: sku_id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Price history
      tags:
      - Price
  /api/v1/refund/initiate:
    post:
      consumes:
      - application/json
      parameters:
      - description: initiate refund
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_refund.InitiateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Initiate refund
      tags:
      - Refund
  /api/v1/refund/rollback:
    post:
      consumes:
      - application/json
      parameters:
      - description: rollback refund
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_refund.RollbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Rollback on cancel
      tags:
      - Refund
  /api/v1/refund/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: refund status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_refund.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Refund status
      tags:
      - Refund
  /api/v1/users/me:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 获取当前用户信息
      tags:
      - User
...
  /orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建订单请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建订单
      tags:
      - Order
  /orders/{id}:
    get:
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询订单
      tags:
      - Order
...
  /payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建支付请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_payment.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建支付
      tags:
      - Payment
  /payments/{id}:
    get:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询支付
      tags:
      - Payment
  /payments/{id}/failed:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付失败
      tags:
      - Payment
...
  /payments/{id}/success:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付成功
      tags:
      - Payment
  /payments/{id}/timeout:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付超时
      tags:
      - Payment
...
swagger: "2.0"